services/orchestrator/requirements.txt
fastapi==0.129.0
uvicorn[standard]==0.40.0
pydantic==2.12.5
pydantic-settings==2.13.0
python-multipart==0.0.22
requests==2.32.5
PyYAML==6.0.3
kubernetes==35.0.0
celery==5.6.2
redis==7.1.1
kombu==5.6.2
cryptography==46.0.5
python-dotenv==1.2.1
